Adds of some changes + port 8080
The server is now using the port 8080.
Creation of packages for tests classes, changes of the javadoc, and some
others changes.
