+++ core-server/src/main/configuration/application.conf
@@ -39,6 +39,11 @@ monitor {
   # period = 10
 }
 
+http.port = 8080
+https.port = 8080
+vertx.http.port = 8080
+vertx.https.port = 8080
+
 # Documentation configuration
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # This property enables or disables the registration of the documentation on the root url.
@@ -48,4 +53,3 @@ documentation.standalone = true
 ## H2 - Memory
 db.h2mem.driver="org.h2.Driver"
 db.h2mem.url="jdbc:h2:mem:h2-mem-it"
-
