+++ core-server/src/main/resources/templates/index.thl.html
@@ -66,7 +66,7 @@
 
 <pre th:text="${liste}">Welcome!</pre>
 
-<li><a type="text/html"  href="orbisserver/ows?service=WPS&amp;version=2.0.0&amp;request=GetCapabilities" th:text="GetCapabilities">GetCapabilities</a></li>
+<li><a type="text/html"  href="orbisserver/wps?service=WPS&amp;version=2.0.0&amp;request=GetCapabilities" th:text="GetCapabilities">GetCapabilities</a></li>
 
 <!-- Scripts -->
 <script src="/libs/jquery.min.js"></script>
